+++ backends/tfhe-cuda-backend/cuda/src/integer/scalar_mul.cu
@@ -28,7 +28,7 @@ void cuda_scalar_multiplication_integer_radix_ciphertext_64_inplace(
   host_integer_scalar_mul_radix<uint64_t>(
       CudaStreams(streams), lwe_array, decomposed_scalar, has_at_least_one_set,
       reinterpret_cast<int_scalar_mul_buffer<uint64_t> *>(mem), bsks,
-      (uint64_t **)(ksks), message_modulus, num_scalars);
+      (uint32_t **)(ksks), message_modulus, num_scalars);
 }
 
 void cleanup_cuda_integer_radix_scalar_mul(CudaStreamsFFI streams,
