+++ backends/tfhe-cuda-backend/cuda/src/integer/scalar_rotate.cu
@@ -27,7 +27,7 @@ void cuda_integer_radix_scalar_rotate_kb_64_inplace(
   host_integer_radix_scalar_rotate_kb_inplace<uint64_t>(
       CudaStreams(streams), lwe_array, n,
       (int_logical_scalar_shift_buffer<uint64_t> *)mem_ptr, bsks,
-      (uint64_t **)(ksks));
+      (uint32_t **)(ksks));
 }
 
 void cleanup_cuda_integer_radix_scalar_rotate(CudaStreamsFFI streams,
