+++ backends/tfhe-cuda-backend/cuda/src/integer/scalar_shifts.cu
@@ -31,7 +31,7 @@ void cuda_integer_radix_logical_scalar_shift_kb_64_inplace(
   host_integer_radix_logical_scalar_shift_kb_inplace<uint64_t>(
       CudaStreams(streams), lwe_array, shift,
       (int_logical_scalar_shift_buffer<uint64_t> *)mem_ptr, bsks,
-      (uint64_t **)(ksks), lwe_array->num_radix_blocks);
+      (uint32_t **)(ksks), lwe_array->num_radix_blocks);
 }
 
 uint64_t scratch_cuda_integer_radix_arithmetic_scalar_shift_kb_64(
@@ -68,7 +68,7 @@ void cuda_integer_radix_arithmetic_scalar_shift_kb_64_inplace(
   host_integer_radix_arithmetic_scalar_shift_kb_inplace<uint64_t>(
       CudaStreams(streams), lwe_array, shift,
       (int_arithmetic_scalar_shift_buffer<uint64_t> *)mem_ptr, bsks,
-      (uint64_t **)(ksks));
+      (uint32_t **)(ksks));
 }
 
 void cleanup_cuda_integer_radix_logical_scalar_shift(CudaStreamsFFI streams,
